+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.home.TaskManager"></activity>
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -20,6 +19,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".ui.home.TaskManager"
+            android:parentActivityName=".MainActivity"/>
     </application>
 
 </manifest>
