+++ app/src/main/res/layout/fragment_home.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -37,6 +38,8 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
         </androidx.viewpager.widget.ViewPager>
+
     </LinearLayout>
-    
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
