+++ app/src/main/res/layout/fragment_task_remaining.xml
@@ -4,10 +4,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".ui.home.FragmentTaskRemaining">
+
 <!--        Отрисовка задач         -->
         <ListView
             android:id="@+id/taskRemaining_list"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"/>
-
 </FrameLayout>
